move rollup-plugin into packages
