+++ package.json
@@ -72,9 +72,9 @@
 		"build:formatters": "rollup -c packages/formatters/rollup.config.js && tsc -p packages/formatters/tsconfig.json --emitDeclarationOnly",
 		"build:node": "rollup -c rollup-node.config.ts",
 		"build:react": "rollup -c rollup-react.config.ts && tsc -p tsconfig-react.json --emitDeclarationOnly",
-		"build:rollup": "rollup -c rollup-rollup.config.ts && tsc -p tsconfig-rollup.json --emitDeclarationOnly",
+		"build:rollup": "rollup -c packages/rollup-plugin/rollup.config.js && tsc -p packages/rollup-plugin/tsconfig.json --emitDeclarationOnly",
 		"build:svelte": "rollup -c rollup-svelte.config.ts && tsc -p tsconfig-svelte.json --emitDeclarationOnly",
-		"build:webpack": "rollup -c packages/webpack/rollup.config.js && tsc -p packages/webpack/tsconfig.json --emitDeclarationOnly",
+		"build:webpack": "rollup -c packages/webpack-plugin/rollup.config.js && tsc -p packages/webpack-plugin/tsconfig.json --emitDeclarationOnly",
 		"generate:json-schema": "mkdir schema && typescript-json-schema ./tsconfig-esm.json GeneratorConfig > ./schema/typesafe-i18n.json",
 		"lint": "eslint src",
 		"lint:fix": "eslint src --fix",
