+++ packages/rollup-plugin/rollup.config.js
@@ -1,26 +1,38 @@
+// @ts-check
+
 import fs from 'fs'
 
 import resolve from '@rollup/plugin-node-resolve'
 import typescript from '@rollup/plugin-typescript'
 import externals from 'rollup-plugin-node-externals'
 import commonjs from '@rollup/plugin-commonjs'
+import path from 'path'
 
-const files = fs.readdirSync('src/')
+const getPath = (file) => path.resolve(__dirname, file)
+
+const files = fs.readdirSync(getPath('src'))
 
 const config = files
 	.filter((file) => file.startsWith('rollup-plugin-typesafe-i18n'))
 	.map((file) => ({
-		input: `src/${file}`,
+		input: getPath(`src/${file}`),
 		output: [
 			{
-				file: `rollup/${file.replace('.ts', '.js')}`,
+				file: getPath(`../../rollup/${file.replace('.ts', '.js')}`),
 				format: 'cjs',
 				sourcemap: true,
 				exports: 'named',
 			},
 		],
 		external: ['typescript', 'chokidar'],
-		plugins: [commonjs(), resolve({ preferBuiltins: true }), externals(), typescript()],
+		plugins: [
+			commonjs(),
+			resolve({ preferBuiltins: true }),
+			externals(),
+			typescript({
+				tsconfig: getPath('./tsconfig.json'),
+			}),
+		],
 	}))
 
 export default config
