+++ packages/rollup-plugin/src/rollup-plugin-typesafe-i18n.ts
