+++ packages/rollup-plugin/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"module": "CommonJS",
+		"outDir": "../../rollup"
+	},
+	"include": ["./src/**/*.ts"]
+}
